+++ test/dev/dev.ts
@@ -6,18 +6,38 @@ import { NO_ADDITIONAL_NODES_PRESET } from '../../src/options/presets/NoCustomNo
 
     let obfuscatedCode: string = JavaScriptObfuscator.obfuscate(
         `
-        var bar = 1;
-        var baz = 2;
         (function(){
-            var bark = bar + baz;
+            function foo () {
+                var a = 1;
+                inner1();
+                var b = 2;
+                function inner1 () {}
+                var c = 3;
+            }
+            function bar () {
+                var a = 1;
+            }
+            function baz () {
+                var a = 1;
+            }
+            function bark () {
+                var a = 1;
+            }
+            function hawk () {
+                var a = 1;
+            }
+            function eagle () {
+                var a = 1;
+            }
         })();
         `,
         {
             ...NO_ADDITIONAL_NODES_PRESET,
             compact: false,
-            identifiersPrefix: 'foo',
-            identifierNamesGenerator: 'mangled',
-            renameGlobals: true
+            stringArray: true,
+            stringArrayThreshold: 1,
+            deadCodeInjection: true,
+            deadCodeInjectionThreshold: 1
         }
     ).getObfuscatedCode();