+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Change Log
 v0.14.3
 ---
 * Fixed https://github.com/javascript-obfuscator/javascript-obfuscator/issues/195
+* Added code preview to `esprima` error messages.
 
 v0.14.2
 ---