+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 Change Log
 ===
+v0.14.3
+---
+* Fixed https://github.com/javascript-obfuscator/javascript-obfuscator/issues/195
+* Added code preview to `esprima` error messages.
+
+v0.14.2
+---
+* Fixed https://github.com/javascript-obfuscator/javascript-obfuscator/issues/181
+
 v0.14.1
 ---
 * Temporary fixed https://github.com/javascript-obfuscator/javascript-obfuscator/issues/181